requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
qdrant-client==1.11.1
sentence-transformers==3.2.1
anthropic==0.40.0
google-generativeai==0.8.3
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
